Rename CLUSTER setting to settings.
